core: conflicts: move requirement types to their file
